requirements.txt
appnope==0.1.0; sys_platform == "darwin" or platform_system == "Darwin" or python_version >= "3.3" and sys_platform == "darwin"
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3; python_version >= "3.3" and sys_platform == "win32" or sys_platform == "win32"
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
flask==1.1.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0; python_version < "3.8"
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
kiwisolver==1.1.0
lxml==4.5.0
markupsafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0; python_version >= "3.3" and sys_platform != "win32" or sys_platform != "win32"
pickleshare==0.7.5
plotly==4.5.4
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0; python_version >= "3.3" and sys_platform != "win32" or sys_platform != "win32" or os_name != "nt" or python_version >= "3.3" and sys_platform != "win32" and (python_version >= "3.3" and sys_platform != "win32" or sys_platform != "win32")
pygments==2.6.1
pyparsing==2.4.6
pyqt5==5.14.1
pyqt5-sip==12.7.1
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227; sys_platform == "win32"
pywinpty==0.5.7; os_name == "nt"
pyzmq==19.0.0
requests==2.23.0
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
send2trash==1.5.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==3.1.0; python_version < "3.8"
